Relocated bounty components outof v3 directory
